+++ spec/unit/cli/scan_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Spandx::Cli::Commands::Scan do
   subject { described_class.new(lockfile, options) }
 
   let(:output) { StringIO.new }
-  let(:lockfile) { nil }
+  let(:lockfile) { '.' }
   let(:options) { {} }
 
   before do
@@ -12,10 +12,33 @@ RSpec.describe Spandx::Cli::Commands::Scan do
       .to_return(status: 200, body: IO.read(fixture_file('spdx/json/licenses.json')))
   end
 
-  it 'executes `scan` command successfully' do
-    subject.execute(output: output)
+  context 'when scanning a directory' do
+    let(:lockfile) { '.' }
+    let(:result) { JSON.parse(output.string) }
 
-    expect(output.string).to eq("OK\n")
+    before do
+      VCR.use_cassette('scan-directory') do
+        subject.execute(output: output)
+      end
+    end
+
+    specify { expect(result['packages'].count).to be(47) }
+    specify { expect(result).to include('version' => '1.0') }
+    specify { expect(result['packages']).to include('name' => 'net-hippie', 'version' => '0.3.2', 'licenses' => ['MIT']) }
+  end
+
+  context 'when recursively scanning a directory' do
+    let(:lockfile) { fixture_file('.') }
+    let(:result) { JSON.parse(output.string) }
+    let(:options) { { 'recursive' => true } }
+
+    before do
+      VCR.use_cassette('scan-directory-recursively') do
+        subject.execute(output: output)
+      end
+    end
+
+    specify { expect(result['packages'].count).to be(26) }
   end
 
   context 'when scanning Gemfile.lock' do