+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow scanning a directory.
+- Allow recursive scanning of a directory.
 
 ## [0.7.0] - 2020-03-11
 ### Changed
@@ -88,7 +91,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Provide ruby API to the latest SPDX catalogue.
 
-[Unreleased]: https://github.com/mokhan/spandx/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/mokhan/spandx/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/mokhan/spandx/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/mokhan/spandx/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/mokhan/spandx/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/mokhan/spandx/compare/v0.4.0...v0.4.1